Ashesi Support Chatbot/CapstoneFinalProduct/requirements.txt
absl-py==0.6.1
APScheduler==3.5.3
astor==0.7.1
attrs==18.2.0
Automat==0.7.0
boto3==1.9.74
botocore==1.12.74
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
Click==7.0
cloudpickle==0.6.1
colorama==0.4.1
colorclass==2.2.0
coloredlogs==10.0
colorhash==1.0.2
ConfigArgParse==0.13.0
constantly==15.1.0
cycler==0.10.0
cymem==1.31.2
cytoolz==0.9.0.1
decorator==4.3.0
dill==0.2.8.2
docopt==0.6.2
docutils==0.14
fakeredis==0.10.3
fbmessenger==5.3.2
Flask==1.0.2
Flask-Cors==3.0.7
Flask-JWT-Simple==0.0.3
future==0.17.1
gast==0.2.1
get==2018.11.19
gevent==1.4.0
greenlet==0.4.15
grpcio==1.17.1
gunicorn==19.9.0
h5py==2.9.0
humanfriendly==4.17
hyperlink==18.0.0
idna==2.8
incremental==17.5.0
itsdangerous==1.1.0
Jinja2==2.10
jmespath==0.9.3
jsonpickle==0.9.6
jsonschema==2.6.0
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
kiwisolver==1.0.1
klein==17.10.0
Markdown==3.0.1
MarkupSafe==1.1.0
matplotlib==2.2.3
mattermostwrapper==2.1
msgpack==0.6.0
msgpack-numpy==0.4.4.2
murmurhash==0.28.0
mysql-connector==2.1.6
networkx==2.2
numpy==1.15.4
packaging==17.1
passlib==1.7.1
pathlib==1.0.1
pg8000==1.13.1
pika==0.11.2
plac==0.9.6
post==2018.11.20
preshed==1.0.1
prompt-toolkit==1.0.14
protobuf==3.6.1
psycopg2-binary==2.7.7
public==2018.11.20
pycparser==2.19
pydot==1.4.1
Pygments==2.3.1
PyHamcrest==1.9.0
PyInquirer==1.0.3
PyJWT==1.7.1
pykwalify==1.6.0
pymongo==3.7.2
pyparsing==2.3.0
pyreadline==2.1
PySocks==1.6.8
python-crfsuite==0.9.6
python-dateutil==2.7.5
python-engineio==3.2.0
python-socketio==2.1.2
python-telegram-bot==10.1.0
pytz==2018.9
PyYAML==3.13
query-string==2018.11.20
rasa-core==0.12.3
rasa-core-sdk==0.12.1
rasa-nlu==0.13.8
redis==2.10.6
regex==2018.11.22
request==2018.11.20
requests==2.21.0
requests-toolbelt==0.8.0
rocketchat-API==0.6.25
ruamel.yaml==0.15.83
s3transfer==0.1.13
scikit-learn==0.19.2
scipy==1.2.0
simplejson==3.16.0
six==1.12.0
sklearn-crfsuite==0.3.6
slackclient==1.3.0
spacy==2.0.12
tabulate==0.8.2
tensorboard==1.10.0
tensorflow==1.10.0
termcolor==1.1.0
terminaltables==3.1.0
thinc==6.10.3
toolz==0.9.0
tqdm==4.29.0
twilio==6.22.1
Twisted==18.9.0
typing==3.6.6
tzlocal==1.5.1
ujson==1.35
urllib3==1.24.1
virtualenv==16.2.0
wcwidth==0.1.7
webexteamssdk==1.1.1
websocket-client==0.54.0
Werkzeug==0.14.1
wrapt==1.10.11
xlwt==1.3.0
zope.interface==4.6.0
